requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.6.2=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2024.6.2=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=41.0.3=py38hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
datasets=2.18.0=pypi_0
dbus=1.13.18=hb2f20db_0
dgl=1.1.2.cu117=py38_0
dill=0.3.8=pypi_0
distlib=0.3.8=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
expat=2.5.0=h6a678d5_0
faiss-gpu=1.7.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
gast=0.4.0=pypi_0
gdown=5.2.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.59.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.10.0=pypi_0
huggingface-hub=0.21.4=pypi_0
hydra=2.5=pypi_0
hydra-core=1.3.2=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
infomap=2.7.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py38h06a4308_0
joblib=1.2.0=py38h06a4308_0
jpeg=9e=h5eee18b_1
keras=2.13.1=pypi_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=16.0.6=pypi_0
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lightning-bolts=0.7.0=pypi_0
lightning-utilities=0.10.1=pypi_0
littleutils=0.2.2=pypi_0
llvmlite=0.41.1=pypi_0
loguru=0.7.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
metis=0.2a5=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
nltk=3.8.1=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numba=0.58.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-dali-cuda120=1.35.0=pypi_0
oauthlib=3.2.2=pypi_0
ogb=1.3.6=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
outdated=0.2.2=pypi_0
packaging=23.1=py38h06a4308_0
pandas=2.0.3=pypi_0
pcre=8.45=h295c915_0
pillow=9.4.0=py38h6a678d5_1
pip=23.2.1=py38h06a4308_0
platformdirs=4.2.0=pypi_0
ply=3.11=py38_0
pooch=1.4.0=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
psutil=5.9.0=py38h5eee18b_0
pyarrow=15.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pymetis=2020.1=py38h69a4267_5
pynndescent=0.5.11=pypi_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=1.9.5=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.0=py38h1128e8f_0
scipy=1.10.1=py38hf6e8229_1
seaborn=0.13.2=pypi_0
sentry-sdk=1.42.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
tensorflow=2.13.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torch-cluster=1.6.3+pt20cu118=pypi_0
torch-geometric=2.4.0=pypi_0
torch-scatter=2.1.2+pt20cu118=pypi_0
torch-sparse=0.6.18+pt20cu118=pypi_0
torch-spline-conv=1.2.2+pt20cu118=pypi_0
torchaudio=2.0.2=py38_cu118
torcheval=0.0.7=pypi_0
torchmetrics=1.3.1=pypi_0
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu118
tornado=6.3.2=py38h5eee18b_0
tqdm=4.65.0=py38hb070fc8_0
transformers=4.38.2=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2023.3=pypi_0
umap-learn=0.5.5=pypi_0
urllib3=1.26.16=py38h06a4308_0
virtualenv=20.25.1=pypi_0
wandb=0.16.4=pypi_0
werkzeug=3.0.0=pypi_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.15.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
